requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
python-dotenv
requests
langchain
langchain-community
langchain-core
faiss-cpu
transformers

sentence-transformers